*** Task Requirement Commit *** B.3.a)
Display a detailed view of the vacation, including all vacation details. This view can also be used to add and update the vacation information

*************************************************

Commit notes for developer reference:

Added label in the UI for the recycler view that displays a list of associated excursions (if any), and a hint for the user on how to add excursions to the vacation at hand. Made improvements to the layout of the UI for readability.