*** Task Requirement Commit *** B.5.D)
Include an alert that the user can set which will trigger on the excursion date, stating the excursion title.

*************************************************

Commit notes for developer reference:

Updated scheduleNotification method in ExcursionDetails to include the title of the excursion as well as including debugging to correct an error that was occurring with passing information for the excursions.