*** Task Requirement Commit *** B.3.B)
"Enter, Edit, and delete vacation information"

The delete functionality was no longer working after updates thus far. the attemptDeleteVacationmethod was updated to inform user why a deletion is not allowed. Logic was added to the Delete Button to call showDeleteConfirmationDialog method, which confirms deletion and enhanced the UI messages for clarity.

*************************************************