*** Task Requirement Commit *** B.5.E)
Include validation that the excursion date is during the associated vacation.

An update to updateVacation in VacationDetails was necessary to ensure that any updates to vacation dates would not be able to circumvent the requirement for excursions to take place during vacations.

Before this was done, a user could update the dates of a vacation and it would not check the dates of the excursion creating the possibility for the excursion to be outside of the vacation dates.

*************************************************